# LeaderboardBackend/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "LeaderboardBackend",
    "version": "v1"
  },
  "paths": {
...
    "/api/leaderboards/{id}": {
      "get": {
        "tags": [
          "Leaderboards"
        ],
        "summary": "Gets a leaderboard by its ID.",
        "operationId": "getLeaderboard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardViewModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/leaderboards/{slug}": {
      "get": {
        "tags": [
          "Leaderboards"
        ],
        "summary": "Gets a leaderboard by its slug. Will not return deleted boards.",
        "operationId": "getLeaderboardBySlug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardViewModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/leaderboards": {
      "get": {
        "tags": [
          "Leaderboards"
        ],
        "summary": "Gets leaderboards. Includes deleted, if specified.",
        "operationId": "listLeaderboards",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "The maximum number of records to return. Fewer records may be returned.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "The zero-based index at which to begin selecting records to return.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/StatusFilter"
                }
              ],
              "default": "Published"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sorts results by a leaderboard's field, tie-breaking with IDs if needed.",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/SortLeaderboardsBy"
                }
              ],
              "description": "Used in GetLeaderboards to sort leaderboards by a field.",
              "default": "Name_Asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardViewModelListView"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/leaderboards": {
      "get": {
        "tags": [
          "Leaderboards"
        ],
        "summary": "Search leaderboards by name or slug.",
        "operationId": "searchLeaderboards",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "The query string. Must not be empty.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The maximum number of records to return. Fewer records may be returned.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "The zero-based index at which to begin selecting records to return.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/StatusFilter"
                }
              ],
              "default": "Published"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardViewModelListView"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/leaderboards": {
      "post": {
        "tags": [
          "Leaderboards"
        ],
        "summary": "Creates a new leaderboard. This request is restricted to Administrators.",
        "operationId": "createLeaderboard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateLeaderboardRequest"
                  }
                ],
                "description": "This request object is sent when creating a `Leaderboard`."
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardViewModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "The requesting `User` is unauthorized to create `Leaderboard`s."
          },
          "409": {
            "description": "A Leaderboard with the specified slug already exists and will be returned in the `conflicting` field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardViewModelConflictDetails"
                }
              }
            }
          },
          "422": {
            "description": "The request contains errors. The following errors can occur: NotEmptyValidator, SlugFormat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/leaderboards/{id}": {
      "delete": {
        "tags": [
          "Leaderboards"
        ],
        "summary": "Deletes a leaderboard. This request is restricted to Administrators.",
        "operationId": "deleteLeaderboard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "The leaderboard does not exist (Not Found) or was already deleted (Already Deleted).\nUse the title field of the response to differentiate between the two cases if necessary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Leaderboards"
        ],
        "summary": "Updates a leaderboard with the specified new fields. This request is restricted to administrators. This operation is atomic; if an error occurs, the leaderboard will not be updated. All fields of the request body are optional but you must specify at least one.",
        "operationId": "updateLeaderboard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateLeaderboardRequest"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The specified slug is already in use by another leaderboard. Returns the conflicting leaderboard.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardViewModelConflictDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateLeaderboardRequest": {
        "required": [
          "name",
          "slug"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "The display name of the `Leaderboard` to create.",
            "example": "Foo Bar"
          },
          "slug": {
            "minLength": 1,
            "type": "string",
            "description": "The URL-scoped unique identifier of the `Leaderboard`.\n\nMust be [2, 80] in length and consist only of alphanumeric characters and hyphens.",
            "example": "foo-bar"
          },
          "info": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "description": "This request object is sent when creating a `Leaderboard`."
      },
...
      "LeaderboardStats": {
        "type": "object",
        "properties": {
          "runCount": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "LeaderboardViewModel": {
        "required": [
          "createdAt",
          "deletedAt",
          "id",
          "info",
          "name",
          "slug",
          "status",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The unique identifier of the `Leaderboard`.\n\nGenerated on creation.",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": "The display name of the `Leaderboard` to create.",
            "example": "Foo Bar"
          },
          "slug": {
            "type": "string",
            "description": "The URL-scoped unique identifier of the `Leaderboard`.\n\nMust be [2, 80] in length and consist only of alphanumeric characters and hyphens.",
            "example": "foo-bar"
          },
          "info": {
            "type": "string",
            "description": "The general information for the Leaderboard.",
            "example": "Timer starts on selecting New Game and ends when the final boss is beaten."
          },
          "createdAt": {
            "type": "string",
            "description": "The time the Leaderboard was created.",
            "format": "date-time",
            "example": "1984-01-01T00:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "The last time the Leaderboard was updated or null.",
            "format": "date-time",
            "nullable": true,
            "example": "1984-01-01T00:00:00Z"
          },
          "deletedAt": {
            "type": "string",
            "description": "The time at which the Leaderboard was deleted, or null if the Leaderboard has not been deleted.",
            "format": "date-time",
            "nullable": true,
            "example": "1984-01-01T00:00:00Z"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Status"
              }
            ]
          },
          "stats": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LeaderboardStats"
              }
            ]
          }
        },
        "additionalProperties": false,
        "description": "Represents a collection of `Leaderboard` entities."
      },
      "LeaderboardViewModelConflictDetails": {
        "required": [
          "conflicting"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "conflicting": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LeaderboardViewModel"
              }
            ],
            "description": "Represents a collection of `Leaderboard` entities.",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "LeaderboardViewModelListView": {
        "required": [
          "data",
          "limitDefault",
          "limitMax",
          "total"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeaderboardViewModel"
            }
          },
          "total": {
            "type": "integer",
            "description": "The total number of records matching the given criteria that\nexist in the database, NOT the total number of records returned.",
            "format": "int64"
          },
          "limitDefault": {
            "type": "integer",
            "description": "The default limit that will be applied for this resource type\nif the client does not specify one in the query string.",
            "format": "int32"
          },
          "limitMax": {
            "type": "integer",
            "description": "The maximum value the client is allowed to specify as a limt for\nendpoints return a paginated list of resources of this type.\nExceeding this value will result in an error.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "SortLeaderboardsBy": {
        "enum": [
          "Name_Asc",
          "Name_Desc",
          "CreatedAt_Asc",
          "CreatedAt_Desc",
          "RunCount_Asc",
          "RunCount_Desc"
        ],
        "type": "string",
        "description": "Used in GetLeaderboards to sort leaderboards by a field."
      },
      "Status": {
        "enum": [
          "Published",
          "Deleted"
        ],
        "type": "string"
      },
      "StatusFilter": {
        "enum": [
          "Published",
          "Deleted",
          "Any"
        ],
        "type": "string"
      },
...
      "UpdateLeaderboardRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "info": {
            "type": "string"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Status"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization using the Bearer scheme",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}